Adds Move-Project76.ps1
Move-Project76 is a quick script to deal with fallout 76 gamepass's disappearing Project76.ini file. It copies the Project76.ini from the installation location to \Documents\My Games\Fallout 76\.  This is an issue for those running Fallout 76 from an install on a drive other than where the OS is installed. Some setup may be required (setting paths if they aren't default).